+++ libs/ui/src/lib/line-chart/line-chart.component.ts
@@ -27,6 +27,7 @@ import {
 import {
   type AnimationsSpec,
   Chart,
+  Decimation,
   Filler,
   LinearScale,
   LineController,
@@ -58,6 +59,7 @@ export class GfLineChartComponent
   @Input() benchmarkLabel = '';
   @Input() colorScheme: ColorScheme = DEFAULT_COLOR_SCHEME;
   @Input() currency: string;
+  @Input() dataDecimation = false;
   @Input() historicalDataItems: LineChartItem[];
   @Input() isAnimated = false;
   @Input() label: string;
@@ -82,6 +84,7 @@ export class GfLineChartComponent
 
   public constructor(private changeDetectorRef: ChangeDetectorRef) {
     Chart.register(
+      Decimation,
       Filler,
       LineController,
       LineElement,
@@ -187,6 +190,9 @@ export class GfLineChartComponent
         this.chart.options.plugins ??= {};
         this.chart.options.plugins.tooltip =
           this.getTooltipPluginConfiguration();
+        this.chart.options.plugins.decimation = this.dataDecimation
+          ? { algorithm: 'lttb', enabled: true }
+          : undefined;
         this.chart.options.animations = this.isAnimated
           ? animations
           : undefined;
@@ -206,6 +212,9 @@ export class GfLineChartComponent
             },
             interaction: { intersect: false, mode: 'index' },
             plugins: {
+              decimation: this.dataDecimation
+                ? { algorithm: 'lttb', enabled: true }
+                : undefined,
               legend: {
                 align: 'start',
                 display: this.showLegend,